third_party/WebKit/LayoutTests/editing/deleting/delete-listitem-003.html

<head>
<style>
  p.listStyle:before {
    content: "\25cf";
    margin-right: 18pt;
  }
</style>
</head>
<body>
<div id="container">
<div contenteditable="true" id="work">
    <p class="listStyle"><span>One</span><br id="one"></p>
    <p class="listStyle"><span>Two</span><br id="two"></p>
    <p class="listStyle"><span>Three</span><br id="three"></p>
</div>
</div>
<script src="../../resources/js-test.js"></script>
<script>
function $(id) { return document.getElementById(id); }

var sample = document.querySelectorAll("p.listStyle")[1];
var range = document.createRange();
range.selectNode(sample.firstChild);
var selection = window.getSelection();
selection.removeAllRanges();
selection.addRange(range);
document.execCommand('Delete');
shouldBeEqualToString('sample.outerHTML', '<p class="listStyle"><br id="two"></p>');

if (window.testRunner)
    $('container').outerHTML = '';
</script>
</body>
