third_party/blink/web_tests/platform/generic/editing/deleting/delete-listitem-003-expected.txt

PASS sample.outerHTML is "<p class=\"listStyle\"><br id=\"two\"></p>"
PASS successfullyParsed is true

TEST COMPLETE
